requirements.txt
Flask
flask-restful
flask-marshmallow
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
flask-redis
Flask-Markdown
marshmallow
marshmallow-sqlalchemy
psycopg2
SQLAlchemy
cryptography
flask_wtf
rq
diffprivlib
pyjwt
requests
beautifulsoup4